+++ inst/include/cheapr/internal/c_core.h
@@ -5,18 +5,20 @@
 // Author: Nick Christofides
 // License: MIT
 
+#ifndef R_NO_REMAP
+#define R_NO_REMAP
+#endif
+
 #include <cpp11.hpp>
 #include <optional>
 #include <type_traits>
+#include <cstring>
+#include <cmath>
 
 #ifdef _MSC_VER
 #define RESTRICT __restrict
 #else
 #define RESTRICT __restrict__
-#endif
-
-#ifndef R_NO_REMAP
-#define R_NO_REMAP
 #endif
 
 #if !defined(OBJSXP) && defined(S4SXP)
@@ -100,9 +102,9 @@ struct r_string_t {
   SEXP value;
   r_string_t() : value{R_BlankString} {}
   // Explicit SEXP -> r_string_t
-  explicit constexpr r_string_t(SEXP x) : value{x} {}
+  explicit r_string_t(SEXP x) : value{x} {}
   // Implicit r_string_t -> SEXP
-  constexpr operator SEXP() const { return value; }
+  operator SEXP() const { return value; }
 };
 
 inline const r_string_t blank_r_string = r_string_t();
@@ -110,9 +112,8 @@ inline const r_string_t blank_r_string = r_string_t();
 // Alias type for SYMSXP
 struct r_symbol_t {
   SEXP value;
-  r_symbol_t() : value{R_MissingArg} {}
-  explicit constexpr r_symbol_t(SEXP x) : value{x} {}
-  constexpr operator SEXP() const { return value; }
+  explicit r_symbol_t(SEXP x) : value{x} {}
+  operator SEXP() const { return value; }
 };
 
 // Alias type for Rcomplex
@@ -172,25 +173,6 @@ inline r_symbol_t double_brackets_sym = static_cast<r_symbol_t>(R_Bracket2Symbol
 inline r_symbol_t brace_sym = static_cast<r_symbol_t>(R_BraceSymbol);
 inline r_symbol_t dots_sym = static_cast<r_symbol_t>(R_DotsSymbol);
 inline r_symbol_t tsp_sym = static_cast<r_symbol_t>(R_TspSymbol);
-inline r_symbol_t name_sym = static_cast<r_symbol_t>(R_NameSymbol);
-inline r_symbol_t base_sym = static_cast<r_symbol_t>(R_BaseSymbol);
-inline r_symbol_t quote_sym = static_cast<r_symbol_t>(R_QuoteSymbol);
-inline r_symbol_t function_sym = static_cast<r_symbol_t>(R_FunctionSymbol);
-inline r_symbol_t namespace_env_sym = static_cast<r_symbol_t>(R_NamespaceEnvSymbol);
-inline r_symbol_t package_sym = static_cast<r_symbol_t>(R_PackageSymbol);
-inline r_symbol_t seeds_sym = static_cast<r_symbol_t>(R_SeedsSymbol);
-inline r_symbol_t na_rm_sym = static_cast<r_symbol_t>(R_NaRmSymbol);
-inline r_symbol_t source_sym = static_cast<r_symbol_t>(R_SourceSymbol);
-inline r_symbol_t mode_sym = static_cast<r_symbol_t>(R_ModeSymbol);
-inline r_symbol_t device_sym = static_cast<r_symbol_t>(R_DeviceSymbol);
-inline r_symbol_t last_value_sym = static_cast<r_symbol_t>(R_LastvalueSymbol);
-inline r_symbol_t spec_sym = static_cast<r_symbol_t>(R_SpecSymbol);
-inline r_symbol_t previous_sym = static_cast<r_symbol_t>(R_PreviousSymbol);
-inline r_symbol_t sort_list_sym = static_cast<r_symbol_t>(R_SortListSymbol);
-inline r_symbol_t eval_sym = static_cast<r_symbol_t>(R_EvalSymbol);
-inline r_symbol_t drop_sym = static_cast<r_symbol_t>(R_DropSymbol);
-inline r_symbol_t missing_arg = static_cast<r_symbol_t>(R_MissingArg);
-inline r_symbol_t unbound_value = static_cast<r_symbol_t>(R_UnboundValue);
 
 }
 
@@ -840,7 +822,11 @@ inline void r_replace(
         }
       }
     } else {
-      std::replace(p_target + start, p_target + start + n, old_val, new_val);
+      for (R_xlen_t i = 0; i < n; ++i) {
+        if (p_target[start + i] == old_val){
+          vec::set_value(p_target, start + i, new_val);
+        }
+      }
     }
   } else {
     for (R_xlen_t i = 0; i < n; ++i) {
@@ -1165,7 +1151,7 @@ inline constexpr bool is_r_na<r_byte_t>(const r_byte_t x){
 }
 
 template<>
-inline constexpr bool is_r_na<r_symbol_t>(const r_symbol_t x){
+inline bool is_r_na<r_symbol_t>(const r_symbol_t x){
   return false;
 }
 
@@ -2184,11 +2170,9 @@ inline void add_attrs(SEXP x, SEXP attrs) {
     const SEXP *p_attributes = list_ptr_ro(attrs);
     const r_string_t *p_names = vector_ptr<const r_string_t>(names);
 
-    r_symbol_t attr_nm;
-
     for (int i = 0; i < Rf_length(names); ++i){
       if (p_names[i] != blank_r_string){
-        attr_nm = r_cast<r_symbol_t>(p_names[i]);
+        r_symbol_t attr_nm = r_cast<r_symbol_t>(p_names[i]);
         if (address(x) == address(p_attributes[i])){
           SEXP dup_attr = SHIELD(Rf_duplicate(p_attributes[i])); ++NP;
           attr::set_attr(x, attr_nm, dup_attr);
@@ -2271,29 +2255,6 @@ inline SEXP deep_copy(SEXP x){
 
 }
 
-
-namespace env {
-inline SEXP get(r_symbol_t sym, SEXP env, bool inherits = true){
-
-  if (TYPEOF(env) != ENVSXP){
-    Rf_error("second argument to '%s' must be an environment", __func__);
-  }
-
-  SEXP val = inherits ? Rf_findVar(sym, env) : Rf_findVarInFrame(env, sym);
-
-  if (val == symbol::missing_arg){
-    Rf_error("arg `sym` cannot be missing");
-  } else if (val == symbol::unbound_value){
-    return r_null;
-  } else if (TYPEOF(val) == PROMSXP){
-    SHIELD(val);
-    val = eval(val, env);
-    YIELD(1);
-  }
-  return val;
-}
-}
-
 // We call R fn`cheapr::set_threads` to make sure the R option is set
 inline void set_threads(uint16_t n){
   uint16_t max_threads = OMP_MAX_THREADS;
@@ -2307,25 +2268,7 @@ inline void set_threads(uint16_t n){
 
 namespace internal {
 
-// A cleaner lambda-based alternative to
-// using the canonical switch(TYPEOF(x))
-//
-// Pass both the SEXP and an auto variable inside a lambda
-// and visit_vector() will assign the auto variable to the
-// correct pointer
-// Then simply deduce its type (via decltype) for further manipulation
-// To be used in a lambda
-// E.g. visit_r_ptr(x, [&](auto p_x) {})
-
-// One must account for the default case via
-// if constexpr (std::is_same_v<T, std::nullptr_t>)
-// Since `NULL` is included in the default case, if you want
-// separate logic to handle this case, just do the below inside the default case
-// if (is_null(x)){
-// ...
-// } else {
-// ...
-// }
+// Do not use
 template <class F>
 decltype(auto) visit_vector(SEXP x, F&& f) {
   switch (CHEAPR_TYPEOF(x)) {
@@ -2341,39 +2284,7 @@ decltype(auto) visit_vector(SEXP x, F&& f) {
   }
 }
 
-// Wrap any callable f, and return a new callable that:
-//   - takes (auto&&... args)
-//   - calls f(args...) inside cpp11::unwind_protect
-
-// Like cpp11::safe but works also  for variadic fns
-template <typename F>
-auto r_safe_impl(F f) {
-  return [f](auto&&... args)
-    -> decltype(f(std::forward<decltype(args)>(args)...)) {
-
-      using result_t = decltype(f(std::forward<decltype(args)>(args)...));
-
-      if constexpr (std::is_void_v<result_t>) {
-        cpp11::unwind_protect([&] {
-          f(std::forward<decltype(args)>(args)...);
-        });
-        // no return; result_t is void
-      } else {
-        return cpp11::unwind_protect([&]() -> result_t {
-          return f(std::forward<decltype(args)>(args)...);
-        });
-      }
-    };
 }
-}
-
-#define r_safe(F)                                                                      \
-internal::r_safe_impl(                                                                 \
-  [&](auto&&... args)                                                                  \
-    -> decltype(F(std::forward<decltype(args)>(args)...)) {                            \
-      return F(std::forward<decltype(args)>(args)...);                                 \
-    }                                                                                  \
-)
 
 } // End of cheapr namespace